+++ include/chipmunk/chipmunk.h
@@ -131,8 +131,8 @@ typedef struct cpSpace cpSpace;
 
 #include "cpSpace.h"
 
-// Munk2D 8.0.0
-#define CP_VERSION_MAJOR 1
+// Munk2D 2.0.0
+#define CP_VERSION_MAJOR 2
 #define CP_VERSION_MINOR 0
 #define CP_VERSION_RELEASE 0